Definition Flags completed (#47)
* Native only classes, virtual functions

ClassFlags::bit6 is only set on native classes that contain native methods and no scripted methods or fields, is presumably an optimization flag

FunctionFlags::bit19 is set on functions with no body, and virtual functions are referenced by name in the Ref_3_4_* classes in reference.script

* Test class flag, quest function flag

ClassFlags::bit7 is defined in Gibbed.RED4.ScriptFormats as IsTestOnly (as is bit6 as IsImportOnly)

FunctionFlags::bit20 is used for methods in ScriptConditionTypeBase and DeviceComponentPS and I've found them called from quest graphs. It seems to be akin to the quest flag used in WitcherScript

FunctionFlags::bit21 is only used in the abstract IWatchdogSystem and a few stray funcs that don't appear to do anything, it reads akin to the test_only class flag but I'm not completely sure

* FieldFlags & FunctionFlags completed

FieldFlags::bit9 is never set on any field and there isn't a test field flag defined yet, so I feel pretty confident in calling it is_test_only.

FunctionFlags::is_safe is only ever set true on Cast() methods that seem to be safe implicit casts, so I've just lightly renamed it.

FunctionFlags::bit19 has been a tricky one. My previous logic was very flawed, I'd only done a shallow investigation. It's only ever set true on non-native methods that either have no body or that don't call any methods where it's not also set true, leading me to conclude that it's something akin to a thread safety flag that assumes functions are safe until they call any non-safe method

* ParameterFlags finished, all flags fleshed out

ParameterFlags::bit3 is used all over the place for reference types, but notably used in Ref_4_2_2_ConstantHandles very clearly as a const parameter, which makes sense as the language has const fields, functions and locals

* Added language support quest function flag

Also added optional semicolon to the end of single-expression function declarations